requirements.txt
pandas>=2.0.0
dash>=2.14.1
plotly>=5.15.0
dash-bootstrap-components>=1.4.1
requests
limesurveyrc2api>=2.0.0
python-dotenv
dash-bootstrap-components
gunicorn>=20.1.0